sphinx: create new onion_router.go file
